Created block design to test the accelerator, created a optimization report, and added files to start the first optimization
The block design is a Microblaze CPU with instruction and data caches, which hinders the frequency. All optimization finding are in the optmization report. The first optimization to be performed is the reduction of the compressor's critical path by added a pipeline stage for indepenent sources - see the hardware design document.